FUSETOOLS2-1160 - partial fix for embedded tutorial between upgrade of extensions
this is straightforward, not very flexible and covers only the case of
users that has not changed the auto-reload on startup embedded
tutorials. Having a proper registry handling all the kinds of tutorials
would be better but given allocated time on VS Code Didact and impact of
other fixes not yet released and as this fix is working in default case
I think we should go with this "mitigation fix" and consider to be in a
releasable state.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>